OAuth 2.0 셋팅

+++ src/main/java/PratiqueGW/com/repository/EpyRepository.java
@@ -3,7 +3,7 @@ package PratiqueGW.com.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import PratiqueGW.com.entity.EpyEntity;
 
-public interface EpyRepository extends JpaRepository<EpyEntity, Integer>{
+public interface EpyRepository extends JpaRepository<EpyEntity, Long>{
 	
 	// 유저 ID가 있는지 체크
 	Boolean existsByEpyId(String epyId);
